+++ CHANGELOG.md
@@ -47,6 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Stop threads created by game from running after soft restart
 - Prevent finalizers from running during soft restart
 - Fixed some static constructors not being tracked properly
+- Added calls to RuntimeInitializeOnLoadMethod attribute methods
 
 ## Changes